feat: remove login and signup screens
